two tooltips to rule them all, i guess
